+++ src/View/Ncurses/NcursesUtils.h
@@ -11,13 +11,13 @@ namespace NcursesUtils
 
     void endNcurses();
 
-    WINDOW *createWindow(uint16_t height, uint16_t width, uint16_t posY, uint16_t posX);
+    WINDOW *createWindow(uint32_t height, uint32_t width, uint32_t posY, uint32_t posX);
 
     void destroyWindow(WINDOW *window);
 
     void initColors();
 
-    const int16_t colorCode(const Color color);
+    const int32_t colorCode(const Color color);
 }
